Устраняет ошибки
