+++ js/data.js
@@ -32,7 +32,7 @@ const EFFECTS = [
     filter: 'invert',
     min: 0,
     max: 100,
-    step: 0.1,
+    step: 1,
     unit: '%',
   },
   {
