+++ js/renderPhotos.js
@@ -1,36 +1,10 @@
-import { renderComments } from './renderComments.js';
+import { renderFullPhoto } from './renderFullPhoto.js';
 
 const miniatureTemplate = document.querySelector('#picture')
   .content
   .querySelector('.picture');
 const miniatureList = document.querySelector('.pictures');
 const miniatureListFragment = document.createDocumentFragment();
-const fullPictureWindow = document.querySelector('.big-picture');
-const fullPictureWindowClose = document.querySelector('.big-picture__cancel');
-
-fullPictureWindowClose.addEventListener('click', () => {
-  fullPictureWindow.classList.add('hidden');
-  document.body.classList.remove('modal-open');
-});
-
-document.addEventListener('keydown', (evt) => {
-  if(evt.key === 'Escape') {
-    fullPictureWindow.classList.add('hidden');
-    document.body.classList.remove('modal-open');
-  }
-});
-
-function getFullPhoto(photo) {
-  fullPictureWindow.classList.remove('hidden');
-  document.body.classList.add('modal-open');
-
-  fullPictureWindow.querySelector('.big-picture__img').querySelector('img').src = photo.url;
-  fullPictureWindow.querySelector('.likes-count').textContent = photo.likes;
-  fullPictureWindow.querySelector('.social__caption').textContent = photo.description;
-  fullPictureWindow.querySelector('.social__comment-total-count').textContent = photo.comments.length;
-
-  renderComments(photo.comments);
-}
 
 const createPhoto = function(photo) {
   const pictureFrame = miniatureTemplate.cloneNode(true);
@@ -40,7 +14,7 @@ const createPhoto = function(photo) {
   pictureFrame.querySelector('.picture__comments').textContent = photo.comments.length;
 
   pictureFrame.onclick = function () {
-    getFullPhoto(photo);
+    renderFullPhoto(photo);
   };
 
   return pictureFrame;
@@ -70,7 +44,7 @@ function onContainerClick(evt) {
   const picture = pictures.find((item) => item.id === +thumbnail.dataset.thumbnailId
   );
 
-  getFullPhoto(picture);
+  renderFullPhoto(picture);
 }
 
 function renderGallery(currentPhotos) {
